refactor(js): CKEditorConfig.js
